feat: flatten package structure
