+++ packages/core/core.ts
@@ -1,0 +1,3 @@
+/** @module ./ */
+export { search } from "./src/search"
+export type { Options, HitDecorator } from "./src/types"
