--- packages/core/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@nosto/search-js-core",
-  "version": "0.1.0",
-  "license": "ISC",
-  "type": "module",
-  "keywords": [],
-  "author": "Nosto",
-  "scripts": {
-    "test": "vitest --run",
-    "lint": "eslint"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
