+++ packages/core/src/applyDecorators.ts
