--- packages/core/src/currencies.ts
@@ -1,4 +1,0 @@
-/** @module currencies */
-/* eslint-disable barrel-files/avoid-barrel-files */
-export { getCurrencyFormatting, type CurrencyConfig } from "./currencies/getCurrencyFormatting"
-export { priceDecorator, type FormattedPrices, type Result } from "./currencies/priceDecorator"
