--- packages/core/src/thumbnails.ts
@@ -1,6 +1,0 @@
-/** @module thumbnails */
-/* eslint-disable barrel-files/avoid-barrel-files */
-export { generateThumbnailUrl, type Props } from "./thumbnails/generateThumbnailUrl"
-export { thumbnailDecorator, nostoThumbnailDecorator, type Config } from "./thumbnails/thumbnailDecorator"
-export { shopifyThumbnailDecorator, type Config as ShopifyConfig } from "./thumbnails/shopifyThumbnailDecorator"
-export type { ThumbnailSize, ShopifySize } from "./thumbnails/types"
