--- packages/core/test/utils/shims.d.ts
@@ -1,3 +1,0 @@
-interface Window {
-  Nosto?: { shopifyScript?: boolean }
-}
