+++ packages/core/vite.config.ts
@@ -1,15 +1,14 @@
 import { resolve } from "path"
 import { defineConfig } from "vitest/config"
-import { baseConfig } from "./vite.config"
-
-const entryPoints = ["src/search.ts", "src/thumbnails.ts", "src/currencies.ts"]
+import { baseConfig } from "../../vite.config"
 
 export default defineConfig({
   ...baseConfig,
   build: {
+    ...baseConfig.build,
     lib: {
       ...baseConfig.build.lib,
-      entry: entryPoints.map(entry => resolve(__dirname, `packages/core/${entry}`)),
+      entry: resolve(import.meta.dirname, "core.ts"),
       fileName: (format, name) => `core/${name}.${format}.js`
     }
   }
