+++ packages/currencies/currencies.ts
@@ -1,0 +1,3 @@
+/** @module currencies */
+export { getCurrencyFormatting, type CurrencyConfig } from "./src/getCurrencyFormatting"
+export { priceDecorator, type FormattedPrices, type Result } from "./src/priceDecorator"
