+++ packages/currencies/src/getCurrencyFormatting.ts
