+++ packages/currencies/test/getCurrencyFormatting.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from "vitest"
-import { getCurrencyFormatting } from "../../src/currencies"
 import { mockNostojs } from "@nosto/nosto-js/testing"
+import { getCurrencyFormatting } from "../currencies"
 
 const currencyFormatsMock = {
   GBP: {
