+++ packages/preact/preact.ts
@@ -1,0 +1,2 @@
+/** @module preact */
+export { usePersonalization } from "./src/hooks/usePersonalization"
