--- packages/preact/src/hooks.ts
@@ -1,2 +1,0 @@
-/** @module preact/hooks */
-export { usePersonalization } from "./hooks/usePersonalization"
