+++ packages/thumbnails/src/generateThumbnailUrl.ts
