+++ packages/thumbnails/src/thumbnailDecorator.ts
