+++ packages/thumbnails/src/types.ts
