+++ packages/thumbnails/src/utils/shims.d.ts
