+++ packages/thumbnails/test/shopifyThumbnailDecorator.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest"
-import { shopifyThumbnailDecorator } from "../../src/thumbnails/shopifyThumbnailDecorator"
+import { shopifyThumbnailDecorator } from "../src/shopifyThumbnailDecorator"
 
 describe("shopifyThumbnailDecorator", () => {
   it("keeps non-shopify URLs as is", () => {
