+++ packages/thumbnails/test/thumbnailDecorator.spec.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it } from "vitest"
-import { thumbnailDecorator } from "../../src/thumbnails/thumbnailDecorator"
+import { thumbnailDecorator } from "../src/thumbnailDecorator"
 import { SearchProduct } from "@nosto/nosto-js/client"
 import { mockSettings } from "@nosto/nosto-js/testing"
