+++ packages/thumbnails/vite.config.ts
@@ -1,0 +1,15 @@
+import { resolve } from "path"
+import { defineConfig } from "vitest/config"
+import { baseConfig } from "../../vite.config"
+
+export default defineConfig({
+  ...baseConfig,
+  build: {
+    ...baseConfig.build,
+    lib: {
+      ...baseConfig.build.lib,
+      entry: resolve(import.meta.dirname, "thumbnails.ts"),
+      fileName: (format, name) => `thumbnails/${name}.${format}.js`
+    }
+  }
+})
